// src/app/models/entities/Entidades.ts
export interface Adjunto {
  idAdjunto?: number;
  nombreArchivo?: string | undefined;
  extension?: string | undefined;
  mimeType?: string | undefined;
  tamaniobytes?: number | undefined;
  ruta?: string | undefined;
  producto?: Producto[] | undefined;
  usuario?: Usuario[] | undefined;
}

export interface AuthCredentials {
  username?: string | undefined;
  password?: string | undefined;
}

export interface DeleteAdjuntoCommand {
  idAdjunto?: number;
}

export interface DeleteHorariosCommand {
  idHorarios?: number;
}

export interface DeleteMaquinaCommand {
  idMaquinas?: number;
}

export interface DeleteUsuarioCommand {
  idUsuario?: number;
}

export interface Detallepedido {
  idDetallePedido?: number;
  idPedidos?: number;
  idProductos?: number;
  cantidad?: number | undefined;
  precioUnitario?: number | undefined;
  subtotal?: number | undefined;
  idPedidosNavigation?: Pedido;
  idProductosNavigation?: Producto;
}

export interface Factura {
  idFactura?: number;
  idUsuario?: number;
  idReservas?: number;
  idPedidos?: number;
  fechaFactura?: Date;
  montoTotal?: number;
  idPedidosNavigation?: Pedido;
  idReservasNavigation?: Reserva;
  idUsuarioNavigation?: Usuario;
}
// ...
export interface Pedido {
  idPedidos?: number;
  idUsuario?: number;
  fechaPedido?: Date;
  total?: number;
  esActivo?: boolean;
  detallepedido?: Detallepedido[] | undefined;
  factura?: Factura[] | undefined;
  idUsuarioNavigation?: Usuario;
}

export interface PedidoCommand {
  idPedidos?: number | undefined;
  idUsuario?: number;
  fechaPedido?: Date;
  total?: number;
  esActivo?: boolean;
}

export interface Persona {
  idPersonas?: number;
  identificacion?: string | undefined;
  nombres?: string | undefined;
  apellidos?: string | undefined;
  numeroCelular?: string | undefined;
  correoElectronico?: string | undefined;
  direccion?: string | undefined;
  sexo?: string | undefined;
  esActivo?: boolean;
  usuario?: Usuario[] | undefined;
}

export interface Producto {
  idProductos?: number;
  idAdjuntos?: number;
  nombreProdcuto?: string | undefined;
  precio?: number;
  stock?: number;
  esActivo?: boolean | undefined;
  detallepedido?: Detallepedido[] | undefined;
  idAdjuntosNavigation?: Adjunto;
}

export interface ProductosCommand {
  idProductos?: number;
  nombreProdcuto?: string | undefined;
  precio?: number;
  stock?: number;
  esActivo?: boolean | undefined;
}

export interface Reserva {
  idReservas?: number;
  idUsuario?: number;
  idMaquinas?: number;
  fechaReserv?: Date;
  total?: number;
  esActivo?: boolean;
  factura?: Factura[] | undefined;
  idMaquinasNavigation?: Maquinas;
  idUsuarioNavigation?: Usuario;
}

export interface ReservasCommand {
  idReservas?: number | undefined;
  idUsuario?: number;
  idMaquinas?: number;
  fechaReserv?: Date;
  total?: number;
  esActivo?: boolean;
}

export interface Rol {
  idRol?: number;
  descripcion?: string | undefined;
  esAdministrador?: boolean;
  esActivo?: boolean;
  usuario?: Usuario[] | undefined;
}
// ...
export interface Usuario {
  idUsuario?: number;
  idAdjunto?: number | undefined;
  idPersonas?: number;
  idRol?: number | undefined;
  nombreUsuario?: string | undefined;
  clave?: string | undefined;
  esActivo?: boolean;
  factura?: Factura[] | undefined;
  idAdjuntoNavigation?: Adjunto;
  idPersonasNavigation?: Persona;
  idRolNavigation?: Rol;
  pedido?: Pedido[] | undefined;
  reserva?: Reserva[] | undefined;
}